Refactor server code into a simplified app-centric version of the MVC architecture
- Moved all the server logic to 'app/'
- Fragmented 'main.py' into the apps 'html_pages', 'proxy' and 'stremio'
- Create the 'manage.py' file to start the server from the root directory
- Update proxy links to omit the port value if it's equal to `None`
- Modify `add_proxy_to_hls_parts`, from 'app/proxy/utils.py', to fill the constant `M3U8_PARTS_HOSTS` and call `update_allowed_hosts` instead of directly modifying `ALLOWED_PROXY_HOSTS`
