+++ app/html_pages/constants.py
@@ -1,0 +1,3 @@
+import os
+
+TEMPLATES_DIR = os.path.join(os.path.dirname(__file__), "templates")
