+++ app/stremio/constants.py
@@ -1,0 +1,10 @@
+MANIFEST = {
+    "id": "org.stremio.huuuuuugo.httpstreams",
+    "version": "0.2.0",
+    "name": "HTTP Streams",
+    "description": "Assista vídeos de alguns dos mais populares sites gratuitos de streaming brasileiros direto no Stremio. Nada de anúcios maliciosos ou torrents abandonados.",
+    "types": ["movie", "series"],
+    "catalogs": [],
+    "resources": ["stream"],
+    "idPrefixes": ["tt"],
+}
